pre_finish

+++ sources/cacher.cpp
@@ -59,7 +59,7 @@ std::vector<double> back(const std::vector<uint>& buf) {
   sum = 0;
   auto time1 = std::chrono::high_resolution_clock::now();
   for(size_t j = 0; j < expcol; j++) {
-    for (size_t i = buf[k]-1; i > -1; i -= sets) {
+    for (int i = buf[k]-1; i > -1; i -= sets) {
       sum += arr[i];
     }
   }
